bestFitModel.py and fitChecker.py
Made two files: fitChecker.py which checks for the rcs of each model fit and it opens all the eps files associated with the obsid.
Creates the bestFitModel.py that finds the best fit rcs model. Does not implement it fully, only looks at which rcs is closest to 1. Does not account for the edge case of two rcs values being equidistant but on opposite sides of 1.0.
Added more columns to the already existing pandas table that was stored in the dataTable.pkl file. Make changes to the excel sheet that the dataframe was made from.
Used the trials.py file to look at different ways for a message box display and act as we wanted it to.