Made fileCounter.py, updated read_log.py
fileCounter counts the number of files inside the qpo_fit folder. Implemented it's logic in read_log.py. The idea is that if there are less than a certain number of files in the qpo_fit folder, it means that the eps files maybe missing, and read_log.py would crash if run on the obsid. Thus the fileChecker logic checks for the number of files in the qpo_fit folder. If these are more than a particular amount, then the code continues to run, otherwise it moves on to the net iteration of the loop, looking at the next obsid.