Created bestModelFileChecker.py
This is used to check whether the osbids contain the qpo_fit folder, and if they do, then it checks for the best_model.txt file. It creates a list and tallies the number of obsids that contain both.
It also tells us the number of obsids that do not contain the qpo_fit folder at all, and ones that contain the qpo_fit folder, but not the best_model.txt file.